Deleted a printf used for debugging and more stuff

+++ PersonDevice/read_temp.ino
@@ -1,6 +1,6 @@
 int sensorPin = 0;
 
-int read_temp()
+float read_temp()
 {
    int reading = analogRead(sensorPin);  
  return ((reading - 242) / 2.0 + 22);
